+++ src/lib/logger/sinks/file.ts
@@ -394,6 +394,8 @@ export class FileSink implements LogSink {
                   disposition: 'lost',
                 }),
           () => describeError(failure),
+          undefined,
+          'FileSink failure handler',
         );
       }
 
@@ -648,6 +650,8 @@ export class FileSink implements LogSink {
                 disposition: 'no_entry',
               }),
         () => describeError(failure),
+        undefined,
+        'FileSink failure handler',
       );
     }
   }
@@ -689,6 +693,8 @@ export class FileSink implements LogSink {
               disposition: 'no_entry',
             }),
       () => describeError(failure),
+      undefined,
+      'FileSink failure handler',
     );
   }
 
@@ -821,6 +827,8 @@ export class FileSink implements LogSink {
               disposition: 'no_entry',
             }),
       () => describeError(failure),
+      undefined,
+      'FileSink failure handler',
     );
   }
 
@@ -868,6 +876,8 @@ export class FileSink implements LogSink {
               disposition: 'lost',
             }),
       () => describeError(failure),
+      undefined,
+      'FileSink failure handler',
     );
   }
 
@@ -933,6 +943,8 @@ export class FileSink implements LogSink {
                 disposition: 'retrying',
               }),
         () => describeError(failure),
+        undefined,
+        'FileSink failure handler',
       );
     }
   }
@@ -1046,6 +1058,7 @@ export class FileSink implements LogSink {
                   () =>
                     `FileSink error writing to ${this.currentLogFile ?? this.logDir}: ${describeError(err)}`,
                   onReported,
+                  'FileSink failure handler',
                 );
 
               if (kind === 'format') {
@@ -1284,6 +1297,8 @@ export class FileSink implements LogSink {
               disposition: 'lost',
             }),
       () => describeError(failure),
+      undefined,
+      'FileSink failure handler',
     );
   }
 
@@ -1629,6 +1644,8 @@ export class FileSink implements LogSink {
                   disposition: 'no_entry',
                 }),
           () => describeError(failure),
+          undefined,
+          'FileSink failure handler',
         );
       });
 
@@ -1846,6 +1863,8 @@ export class FileSink implements LogSink {
                   disposition: 'no_entry',
                 }),
           () => describeError(failure),
+          undefined,
+          'FileSink failure handler',
         );
       }
 
@@ -1921,6 +1940,8 @@ export class FileSink implements LogSink {
               disposition: 'no_entry',
             }),
       () => describeError(failure),
+      undefined,
+      'FileSink failure handler',
     );
 
     return candidate;
@@ -1958,6 +1979,7 @@ export class FileSink implements LogSink {
         () =>
           `FileSink error rendering an entry for ${this.currentLogFile ?? this.logDir}: ${describeError(failure)}`,
         onReported,
+        'FileSink failure handler',
       );
     });
   }